Paris to prague (#56)
* Worked around Erigon syncing from scratch issues

* Duplicate sync removal

* Testing if it's possible to remove retries during node restart from scratch

* Added dynamically switchable Engine API Client Between Paris and Prague

* Generalized Maru and Besu creation

* Fixed besu genesis

* Fixed E2E tests

* Made switches based on timestamps instead of block numbers. E2E tests are to be fixed

* Implemented switch from Paris to Prague

* Spotless

* Minor fix

* Removed deposit contract from the genesis file

* Removed prague and paris client flexibility

* Removed some leftovers

* Removed misleading comment

* Removed misleading comment

* Clarified a comment

* Reusing some of the mappers

* Removed misleading comment

* Added some tests and made some other improvements

* Moved some logic from OmniProtocolFactory to protocol specific factories

* Logging improvements

+++ maru/docker/initialization/init.sh
@@ -1,15 +1,20 @@
 #!/bin/zsh
-echo "Initialization of timestamp in genesis files for Besu and Geth."
-date
-cd initialization
-cp -T "genesis-besu.json.template" "genesis-besu.json"
-cp -T "genesis-geth.json.template" "genesis-geth.json"
-cp -T "genesis-nethermind.json.template" "genesis-nethermind.json"
+echo "Checking if genesis files already exist."
 
-merge_timestamp=$(($(date +%s) + 60))
-sed -i "s/^    \"shanghaiTime\": .*/    "\"shanghaiTime\"": $merge_timestamp,/" genesis-besu.json
-sed -i "s/^    \"cancunTime\": .*/    "\"cancunTime\"": $merge_timestamp,/" genesis-besu.json
-sed -i "s/^    \"shanghaiTime\": .*/    "\"shanghaiTime\"": $merge_timestamp,/" genesis-geth.json
-sed -i "s/^    \"cancunTime\": .*/    "\"cancunTime\"": $merge_timestamp,/" genesis-geth.json
-merge_timestamp_hex=$(printf "0x%x" $merge_timestamp)
-sed -i "s/%SWITCH_TIME%/$merge_timestamp_hex/g" genesis-nethermind.json
+if [[ ! -f "initialization/genesis-besu.json" && ! -f "initialization/genesis-geth.json" && ! -f "initialization/genesis-nethermind.json" ]]; then
+    echo "Initialization of timestamp in genesis files for Besu, Geth, and Nethermind."
+    date
+    cd initialization || exit
+    cp -T "genesis-besu.json.template" "genesis-besu.json"
+    cp -T "genesis-geth.json.template" "genesis-geth.json"
+    cp -T "genesis-nethermind.json.template" "genesis-nethermind.json"
+
+    merge_timestamp=$(($(date +%s) + 60))
+    echo "Timestamp: $merge_timestamp"
+    sed -i "s/%SWITCH_TIME%/$merge_timestamp/g" genesis-besu.json
+    sed -i "s/%SWITCH_TIME%/$merge_timestamp/g" genesis-geth.json
+    merge_timestamp_hex=$(printf "0x%x" $merge_timestamp)
+    sed -i "s/%SWITCH_TIME%/$merge_timestamp_hex/g" genesis-nethermind.json
+else
+    echo "Genesis files already exist. Initialization skipped."
+fi
